refactor dokku_clone module
reuse functionality from dokku_app
